refactor: apply grid patches via store produce
